Node-Manager/notification-manager/requirements.txt
confluent-kafka==2.1.0
python-decouple==3.8